refactor(direct-control): extract session class owner
Refs: openspec/changes/civ7-support-direct-control-modularization

Moves Civ7DirectControlSession and its private pending-request lifecycle into src/session/session.ts while keeping the public facade export and package-level query/execute wrappers in index.ts.

This preserves socket connection fallback, LSQ state queries, framed command execution, request timeout/close/error classification, and public facade exports. index.ts still owns package-level query/execute wrappers, session creation/close for public wrappers, reconnect orchestration, lifecycle composition, telemetry/correlation contracts, public procedure schemas, semantic CLI projection, AI ingestion, hotseat runtime proof, and Effect/oRPC procedure-core work. Task 2.9.4 matrix-row acceptance remains pending.

Validation:
- bun run --cwd packages/civ7-direct-control test -- session.test.ts public-api.test.ts
- bun run --cwd packages/civ7-direct-control check
- bun run --cwd packages/civ7-direct-control test
- bun run --cwd packages/civ7-direct-control build
- bun run openspec -- validate civ7-support-direct-control-modularization --strict
- bun run resources:status
- bun run check:cli
- bun run test:cli:play
- bun run --cwd apps/mapgen-studio check
- git diff --check
- git diff --cached --check
- ownership scan: Civ7DirectControlSession and PendingCiv7TunerRequest are owned by src/session/session.ts
- no-cycle scan: session modules do not import executable facade values

Local package/source relocation proof only; no runtime, live-game, AI-on-hotseat, telemetry, semantic CLI, or procedure-core proof claimed.
